Send document.title as title to Chartbeat
Chartbeat expects the `document.title` (`props.title` on Segment `page`)
to be the default value for title. The current behavior first checks
for the page name and category to send as title, falling back to
`props.title` if both name and category are not present.

This introduces a setting to allow current users relying on page
name and category populating as title in Chartbeat to continue
this legacy behavior, while future users will default to the
expected behavior. Current users will also be able to opt into
the correct behavior by way of this setting `sendTitleAndCategory`.